Fix incorrect payload access when retrieving aliases
There were two instances in the code where the current aliases are
fetched from the OpenSearch cluster. The method call will return a
dictionary. However, the code was assuming a Response and thus called
`#body` on it. Since it's a dictionary, we can iterate the values
directly.

The fallout from this was that a reindex would finish (new index
created, default alias pointed at it) but the old indices wouldn't be
deleted. This meant wasted disk space but also, more annoyingly, the
same alias pointed at multiple indices. As it turns out, this is an
option and basically means that you can query against a single alias
which will then forward it to *multiple* indices. For us this meant we
were getting lots of duplicate results.

A second occurence of the same bug can be found in the `/catalog`
endpoint. In this case, it only meant that it didn't work but no
repercussions for the querying.